requirements.txt
streamlit
google-generativeai
PIL
pillow
python-dotenv